add tests and stories for Device component
