+++ src/components/resources/Device/index.js
@@ -1,0 +1,3 @@
+import Device from './Device';
+
+export default Device;
