+++ src/fixtures/dstu2/resources/device/example2.json
@@ -1,0 +1,58 @@
+{
+  "resourceType": "Device",
+  "id": "example",
+  "text": {
+    "fhir_comments": [
+      "  text>\n    <status value=\"generated\"/>\n    <div xmlns=\"http://www.w3.org/1999/xhtml\">\n      <p>example</p>\n    </div>\n  </text  "
+    ],
+    "status": "generated",
+    "div": "<div><p><b>Generated Narrative with Details</b></p><p><b>id</b>: example</p><p><b>identifier</b>: 345675, Serial Number = AMID-342135-8464</p><p><b>type</b>: ECG <span>(Details : {SNOMED CT code '86184003' = '86184003', given as 'Electrocardiographic monitor and recorder'})</span></p><p><b>note</b>: QA Checked</p><p><b>status</b>: available</p><p><b>manufacturer</b>: Acme Devices, Inc</p><p><b>model</b>: AB 45-J</p><p><b>lotNumber</b>: 43453424</p><p><b>contact</b>: ph: ext 4352</p></div>"
+  },
+  "identifier": [
+    {
+      "system": "http://goodcare.org/devices/id",
+      "value": "345675"
+    },
+    {
+      "type": {
+        "coding": [
+          {
+            "system": "http://hl7.org/fhir/identifier-type",
+            "code": "SNO"
+          }
+        ],
+        "text": "Serial Number"
+      },
+      "value": "AMID-342135-8464"
+    }
+  ],
+  "type": {
+    "coding": [
+      {
+        "system": "http://snomed.info/sct",
+        "code": "86184003",
+        "display": "Electrocardiographic monitor and recorder"
+      }
+    ],
+    "text": "ECG"
+  },
+  "note": [
+    {
+      "authorReference": {
+        "reference": "Practitioner/xcda-author"
+      },
+      "time": "2015-06-28T14:03:32+10:00",
+      "text": "QA Checked"
+    }
+  ],
+  "status": "available",
+  "manufacturer": "Acme Devices, Inc",
+  "model": "AB 45-J",
+  "lotNumber": "43453424",
+  "contact": [
+    {
+      "system": "phone",
+      "value": "ext 4352"
+    }
+  ]
+}
